+++ data/io_test.go
@@ -1,4 +1,4 @@
-package ft
+package data
 
 import (
 	"os"
