+++ data/logo.go
@@ -1,0 +1,4 @@
+package data
+
+// default value for logo
+var Logo string = "fastTravelCLI"
