+++ ft/helpdocs.go
@@ -1,0 +1,281 @@
+package ft
+
+import "fmt"
+
+// TODO: add helpdocs to man pages
+// Help docs
+var HelpLineStrFormat = fmt.Sprintf("  %%-%ds %%s\n", FindUsageMaxLen(HelpUsageMappings))
+
+const HelpDescExamples = `
+    Examples: 
+	  ft 						  →	Shows bookmarks for fuzzy selection (requires tree and fzf)
+	  ft projects                 → cd into the directory saved as 'projects'
+	  ft -set docs=~/Documents    → set key 'docs' to '~/Documents'
+	  ft -rn -y old new           → rename 'old' key to 'new' and skip confirmation prompt
+`
+
+var HelpUsageMappings = map[string]string{
+	"key":     "ft <key>",
+	"set":     "ft -set [-y] <key>=[path]",
+	"ls":      "ft ls",
+	"rm":      "ft -rm [-y] <key>",
+	"rn":      "ft -rn [-y] <key> <new key>",
+	"edit":    "ft -edit [-y] <path> <new dir name>",
+	"]":       "ft ]",
+	"[":       "ft [",
+	"hist":    "ft -hist",
+	"is":      "ft -is",
+	"version": "ft -version, -v",
+	"update":  "ft -update, -u [version]",
+	"help":    "ft -help, -h [command]",
+	"fzf":     "ft",
+	"fzfc":    "ft -f [key or path]",
+	"fzfa":    "ft -fa [key or path]",
+}
+
+var CmdDesc = []map[string]string{
+	{
+		HelpUsageMappings["key"]: "Change directory to a saved key",
+	},
+	{
+		HelpUsageMappings["set"]: "Save a key to a path (defaults to CWD if no path)",
+	},
+	{
+		HelpUsageMappings["ls"]: "List all saved key-path pairs",
+	},
+	{
+		HelpUsageMappings["rm"]: "Remove a saved key",
+	},
+	{
+		HelpUsageMappings["rn"]: "Renames an existing key",
+	},
+	{
+		HelpUsageMappings["edit"]: "Updates a given directory to it's new name for any key assigned to it or a child directory",
+	},
+	{
+		HelpUsageMappings["]"]: "Navigate history forwards",
+	},
+	{
+		HelpUsageMappings["["]: "Navigate history backwards",
+	},
+	{
+		HelpUsageMappings["hist"]: "Show directory history with fzf",
+	},
+	{
+		HelpUsageMappings["is"]: "Show the key associated with CWD",
+	},
+	{
+		HelpUsageMappings["version"]: "Show current version",
+	},
+	{
+		HelpUsageMappings["update"]: "Update fastTravelCLI, optionally specify version or 'nightly', defaults to latest stable release",
+	},
+	{
+		HelpUsageMappings["help"]: "Show this help message",
+	},
+	{
+		HelpUsageMappings["fzf"]: "Show bookmarks with fzf",
+	},
+	{
+		HelpUsageMappings["fzfc"]: "Show current level project directories with fzf",
+	},
+	{
+		HelpUsageMappings["fzfa"]: "Show all project directories with fzf",
+	},
+}
+
+// Detailed help docs
+var DetailedCmdDescriptions = [][]string{
+	{HelpUsageMappings["key"],
+		`Description:
+  Changes the current directory to the one associated with the given key.
+
+Accepted Formats for <key>:
+  - A valid saved key
+  - A key/subdir format (e.g., pr/docs/work)
+  - A relative path (e.g., ./mydir)
+  - An absolute path (e.g., /home/user/mydir)
+
+Flags:
+  -h, -help  Show this help message.
+
+Examples:
+  ft projects         # cd into the path saved as "projects"
+  ft dev/mytool       # cd into a subdirectory under the "dev" key`},
+
+	{HelpUsageMappings["set"],
+		`Description:
+  Saves a key to a directory path.
+  If no path is specified, the current working directory is used.
+
+Arguments:
+  key         The name to assign the path to.
+  path        (Optional) The path to associate with the key.
+
+Flags:
+  -y         Skip confirmation prompt.
+  -h, -help  Show this help message.
+
+Examples:
+  ft -set docs                        # Set "docs" to the current working directory
+  ft -set code=~/Projects             # Set "code" to a specific path
+  ft -set os=~/Projects/opensource    # Set "os" to specified path without confirmation prompt`},
+	{HelpUsageMappings["ls"],
+		`Description:
+  Lists all saved key-path pairs.
+
+Flags:
+  -h, -help  Show this help message.
+  `},
+	{HelpUsageMappings["rm"],
+		`Description:
+  Deletes a saved key.
+
+Arguments:
+  key         The key you wish to remove.
+
+Flags:
+  -y    Skip confirmation prompt.
+  -h, -help  Show this help message.
+
+Examples:
+  ft -rm archive       # Prompts for confirmation
+  ft -rm archive -y    # Removes without confirmation`},
+	{HelpUsageMappings["rn"],
+		`Description:
+  Renames an existing key to a new one.
+
+Arguments:
+  key         The key you wish to rename.
+  new key     The name you wish to rename the key to.
+
+Flags:
+  -y    Skip confirmation prompt.
+  -h, -help  Show this help message.
+
+Examples:
+  ft -rn old new       # "old" is renamed to "new"
+  ft -rn old new -y    # "old" is renamed to "new" without confirmation`},
+	{HelpUsageMappings["edit"],
+		`Description:
+  Renames a directory (and all child directories) associated with a key or path.
+  Updates any keys pointing to subdirectories of the renamed folder.
+
+Arguments:
+  path          The path to the directory that needs to be renamed.
+  new dir name  The new name of the directory.
+
+Accepted Formats for <path>:
+  - A relative path (e.g., ./mydir)
+  - An absolute path (e.g., /home/user/mydir)
+  - A key/subdir format (e.g., projects/docs)
+
+Flags:
+  -y    Skip confirmation prompt.
+  -h, -help  Show this help message.
+
+Examples:
+  ft -edit ./api api-v2
+  ft -edit projects/docs documents`},
+	{HelpUsageMappings["]"],
+		`Description:
+  Navigate forward in directory history.
+Flags:
+  -h, -help  Show this help message.`},
+	{HelpUsageMappings["["],
+		`Description:
+  Navigate backward in directory history.
+
+Flags:
+  -h, -help  Show this help message.`},
+	{HelpUsageMappings["hist"],
+		`Description:
+  Shows an interactive history for fuzzy selection.
+
+Flags:
+  -h, -help  Show this help message.
+
+Notes:
+  Requires 'fzf' and 'tree' to be installed.`},
+	{HelpUsageMappings["version"],
+		`Description:
+  Displays the current version of fastTravelCLI.
+
+Flags:
+  -h, -help  Show this help message.`},
+	{HelpUsageMappings["is"],
+		`Description:
+  Displays the key associated with the current working directory, if one exists.
+
+Flags:
+  -h, -help  Show this help message.`},
+	{HelpUsageMappings["update"],
+		`Description:
+  Updates fastTravelCLI to the latest version if version is unspecified.
+  If a version is specified, attempts to install that version.
+  Use "nightly" to install the latest pre-release build.
+
+Arguments:
+  version         The version to update to. Must be either 'nightly' or have format v.n.n.n.
+
+Flags:
+  -h, -help  Show this help message.
+
+Examples:
+  ft -update
+  ft -update v.0.2.1
+  ft -update latest`},
+	{"This is a standard directory navigation command.", "ft can replace cd entirely as it inherits cd's commands."},
+
+	{HelpUsageMappings["fzfc"],
+		`Description:
+  Shows immediate child directories in current or optionally provided project directory for fuzzy selection.
+
+Flags:
+  -h, -help  Show this help message.
+
+Examples:
+	ft -f
+	ft -f somekey
+	ft -f some/dir
+
+Notes:
+  Requires 'fzf' and 'tree' to be installed.`},
+	{HelpUsageMappings["fzfa"],
+		`Description:
+  Shows all child directories in current or optionally provided project directory for fuzzy selection.
+
+Flags:
+  -h, -help  Show this help message.
+
+Examples:
+	ft -fa
+	ft -fa somekey
+	ft -fa some/dir
+
+Notes:
+  Requires 'fzf' and 'tree' to be installed.`},
+}
+
+const HelpCmdMsg = "Use ft -help information on other commands"
+
+var DetailedCmdDescMapping = map[string]string{
+	"_":        CreateCmdHelpDoc(DetailedCmdDescriptions[0]),
+	"-set":     CreateCmdHelpDoc(DetailedCmdDescriptions[1]),
+	"-ls":      CreateCmdHelpDoc(DetailedCmdDescriptions[2]),
+	"-rm":      CreateCmdHelpDoc(DetailedCmdDescriptions[3]),
+	"-rn":      CreateCmdHelpDoc(DetailedCmdDescriptions[4]),
+	"-edit":    CreateCmdHelpDoc(DetailedCmdDescriptions[5]),
+	"-]":       CreateCmdHelpDoc(DetailedCmdDescriptions[6]),
+	"-[":       CreateCmdHelpDoc(DetailedCmdDescriptions[7]),
+	"-hist":    CreateCmdHelpDoc(DetailedCmdDescriptions[8]),
+	"-version": CreateCmdHelpDoc(DetailedCmdDescriptions[9]),
+	"-v":       CreateCmdHelpDoc(DetailedCmdDescriptions[9]),
+	"-is":      CreateCmdHelpDoc(DetailedCmdDescriptions[10]),
+	"-update":  CreateCmdHelpDoc(DetailedCmdDescriptions[11]),
+	"-u":       CreateCmdHelpDoc(DetailedCmdDescriptions[11]),
+	"--":       CreateCmdHelpDoc(DetailedCmdDescriptions[12]),
+	"-..":      CreateCmdHelpDoc(DetailedCmdDescriptions[12]),
+	"-fzfc":    CreateCmdHelpDoc(DetailedCmdDescriptions[13]),
+	"-fzfa":    CreateCmdHelpDoc(DetailedCmdDescriptions[14]),
+}
