+++ ft/query.go
