+++ ft/settings.go
@@ -6,3 +6,14 @@ package ft
 
 // - toggle for history across active sessions (think tmux workflows)
 // - auto updates (or check for updates)
+
+
+
+
+type Settings struct {
+	cascadeOrder []string
+}
+
+func NewSettings() *Settings {
+	return &Settings{}
+}
